fix: breadcrumb factory to separate class
